Fin tme1

+++ Code/test.py
@@ -4,14 +4,24 @@ import ParserCACM as pars
 import TextRepresenter as txtrep
 import Index
 import pickle as pkl
+from pathlib import Path
 
-pars = pars.ParserCACM()
-trep = txtrep.PorterStemmer()
-data = "../cacm/cacm.txt"
-ind = Index.IndexOnFile("test", pars, trep, data)
 path = "../test"
 
-ind.indexation(path)
-with open(path+"/index.pkl","wb") as f:
-    pkl.dump(ind, f)
+
+if Path(path+"/index.pkl").is_file() :
+    with open(path+"/index.pkl","rb") as f:
+        ind = pkl.load(f)
+else :
+    pars = pars.ParserCACM()
+    trep = txtrep.PorterStemmer()
+    data = "../cacm/cacm.txt"
+    ind = Index.IndexOnFile("test", pars, trep)
+    pars.initFile(data)
+    ind.cheatIndexation(path)
+    with open(path+"/index.pkl","wb") as f:
+        pkl.dump(ind, f)
+
+
+
 #print(pkl.dumps(ind, protocol=0))
